requirements.txt
Flask==2.2.2
openai
load_dotenv